+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { createLinkPlugin, LINK_TYPE } from 'wix-rich-content-plugin-link';
 import { createLineSpacingPlugin, LINE_SPACING_TYPE } from 'wix-rich-content-plugin-line-spacing';
 import { createHashtagPlugin, HASHTAG_TYPE } from 'wix-rich-content-plugin-hashtag';
@@ -21,7 +22,6 @@ import {
 import { createMapPlugin, MAP_TYPE } from 'wix-rich-content-plugin-map';
 import { createFileUploadPlugin, FILE_UPLOAD_TYPE } from 'wix-rich-content-plugin-file-upload';
 import { createTextColorPlugin, TEXT_COLOR_TYPE } from 'wix-rich-content-plugin-text-color';
-import React from 'react';
 import Highlighter from 'react-highlight-words';
 import casual from 'casual-browserify';
 
@@ -148,7 +148,8 @@ const uiSettings = {
 
 export const config = {
   [GALLERY_TYPE]: {
-    scrollingElement: () => typeof window !== 'undefined' && document.getElementsByClassName('editor-example')[0],
+    scrollingElement: () =>
+      typeof window !== 'undefined' && document.getElementsByClassName('editor-example')[0],
   },
   [IMAGE_TYPE]: {
     imageEditorWixSettings: {
@@ -177,7 +178,10 @@ export const config = {
   [EXTERNAL_MENTIONS_TYPE]: {
     repositionSuggestions: true,
     visibleItemsBeforeOverflow: 5,
+    popoverComponent: <div />,
+    handleDropdownOpen: () => console.log('mentions dropdown opened'),
     onMentionClick: mention => console.log({ mention }),
+    handleDropdownClose: () => console.log('mentions dropdown closed'),
     getMentions: searchQuery =>
       new Promise(resolve =>
         setTimeout(